# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sport API",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/teams/information": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Lấy thông tin về các mùa giải",
                "description": "Truy vấn thông tin về các mùa giải của giải đấu bóng đá",
                "operationId": "getSeasons",
                "responses": {
                    "200": {
                        "description": "Dữ liệu mùa giải đã được lấy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Dữ liệu mùa giải",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "Thông điệp trả về",
                                            "type": "string",
                                            "example": "Seasons data retrieved successfully"
                                        },
                                        "status": {
                                            "description": "Mã trạng thái HTTP",
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi nội bộ máy chủ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Dữ liệu rỗng",
                                            "type": "null"
                                        },
                                        "message": {
                                            "description": "Thông điệp lỗi",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        },
                                        "status": {
                                            "description": "Mã trạng thái HTTP",
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/teams/statistics": {
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Lấy thống kê về một đội bóng",
                "description": "Truy vấn thông tin thống kê về một đội bóng trong một mùa giải cụ thể",
                "operationId": "getStatistics",
                "parameters": [
                    {
                        "name": "team",
                        "in": "query",
                        "description": "ID của đội bóng",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Ngày thống kê",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dữ liệu thống kê đã được lấy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Dữ liệu thống kê",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "Thông điệp trả về",
                                            "type": "string",
                                            "example": "Statistics data retrieved successfully"
                                        },
                                        "status": {
                                            "description": "Mã trạng thái HTTP",
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi nội bộ máy chủ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Thông điệp lỗi",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Thông điệp lỗi",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        },
                                        "status": {
                                            "description": "Mã trạng thái HTTP",
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/timezones": {
            "get": {
                "tags": [
                    "Timezone"
                ],
                "summary": "Lấy dữ liệu múi giờ",
                "description": "Lấy dữ liệu múi giờ từ API thứ ba",
                "operationId": "getIndex",
                "responses": {
                    "200": {
                        "description": "Dữ liệu múi giờ đã được lấy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Dữ liệu múi giờ",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "Thông điệp trả về",
                                            "type": "string",
                                            "example": "Timezone data retrieved successfully"
                                        },
                                        "status": {
                                            "description": "Mã trạng thái HTTP",
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi nội bộ máy chủ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Dữ liệu rỗng",
                                            "type": "null"
                                        },
                                        "message": {
                                            "description": "Thông điệp lỗi",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        },
                                        "status": {
                                            "description": "Mã trạng thái HTTP",
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}